pyproject.toml
[project]
name = "ai-tech-catchup-agent"
version = "1.0.0"
description = "最新AI情報キャッチアップ用AI Agent"
authors = [
    {name = "Yusuke Sakai"}
]
readme = "README.md"
requires-python = ">=3.8.1"
dependencies = [
    "anthropic>=0.8.0",
    "requests>=2.31.0",
    "feedparser>=6.0.10",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.24.0",
    "lxml>=4.9.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
